Cap3.src/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
filelock==3.25.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
peft==0.13.2
pillow==12.0.0
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
Pygments==2.20.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.0
rich==14.3.3
safetensors==0.7.0
scipy==1.16.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
trl==0.15.2
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
